refactor(Astro): Update tsconfig paths, remove unused Layout component, and refactor page structure to improve component imports
