+++ examples/astro/src/components/Layout.astro
