+++ examples/astro/src/components/content-types/Activity.tsx
