+++ examples/astro/src/components/content-types/Banner.tsx
