+++ examples/astro/src/components/content-types/BannerCarousel.tsx
