+++ examples/astro/src/components/content-types/CalendarEvent.tsx
