+++ examples/astro/src/components/content-types/CallToAction.tsx
