+++ examples/astro/src/components/content-types/CategoryFilter.tsx
