+++ examples/astro/src/components/content-types/Empty.tsx
