+++ examples/astro/src/components/content-types/Image.tsx
