+++ examples/astro/src/components/content-types/PageForm.tsx
@@ -1,6 +1,6 @@
 import type { DotCMSBasicContentlet } from "@dotcms/types";
 
-import { useIsEditMode } from "@react/hooks/isEditMode";
+import { useIsEditMode } from "src/hooks/isEditMode";
 
 interface PageFormProps extends DotCMSBasicContentlet {
   formType: string;
