+++ examples/astro/src/components/content-types/Product.tsx
