+++ examples/astro/src/components/content-types/StoreProductList.tsx
