+++ examples/astro/src/components/content-types/WebPageContent.tsx
