+++ examples/astro/src/components/content-types/index.tsx
