+++ examples/astro/src/components/react/RecommendedCard.tsx
