+++ examples/astro/src/components/react/notFound.tsx
