+++ examples/astro/src/hooks/isEditMode.ts
