+++ examples/astro/src/react/layout/footer/Footer.tsx
