+++ examples/astro/src/react/layout/footer/components/Blogs.tsx
