+++ examples/astro/src/utils/client.ts
@@ -1,5 +1,11 @@
 import { createDotCMSClient } from "@dotcms/client/next";
-import type { DotCMSNavigationItem, DotCMSPageAsset } from "@dotcms/types";
+import type { DotCMSComposedPageResponse, DotCMSNavigationItem, DotCMSPageAsset } from "@dotcms/types";
+import {
+  blogQuery,
+  destinationQuery,
+  fragmentNav,
+  navigationQuery,
+} from "./queries";
 
 export const dotCMSClient = createDotCMSClient({
   authToken: import.meta.env.PUBLIC_DOTCMS_AUTH_TOKEN,
@@ -44,65 +50,3 @@ export const getDotCMSPage = async (
     return null;
   }
 };
-
-export const blogQuery = `
-    search(query: "+contenttype:Blog +live:true", limit: 3) {
-        title
-        identifier
-        ... on Blog {
-            inode
-            image {
-                fileName
-            }
-            urlMap
-            modDate
-            urlTitle
-            teaser
-            author {
-                firstName
-                lastName
-                inode
-            }
-        }
-    }
-`;
-
-export const destinationQuery = `
-    search(query: "+contenttype:Destination +live:true", limit: 3) {
-        title
-        identifier
-        ... on Destination {
-                inode
-                image {
-                fileName
-                }
-                urlMap
-                modDate
-                url
-        }
-    }
-`;
-
-export const navigationQuery = `
-DotNavigation(uri: "/", depth: 2) {
-    ...NavProps
-    children {
-        ...NavProps
-    }
-}
-`;
-
-export const fragmentNav = `
-fragment NavProps on DotNavigation {
-    code
-    folder
-    hash
-    host
-    href
-    languageId
-    order
-    target
-    title
-    type
-}
-`;
